fix(dashboard): remove all pages -> wrappers (finally!) code cleanup + better organization of concerns, fix sidebar displaying deleted tools/policies
