--- packages/apps/app-dashboard/src/pages/developer-dashboard/app/index.ts
@@ -1,1 +1,0 @@
-export { default as AppsPage } from './AppsPage';
