--- packages/apps/app-dashboard/src/pages/developer-dashboard/policy/index.ts
@@ -1,1 +1,0 @@
-export { default as PoliciesPage } from './PoliciesPage';
