--- packages/apps/app-dashboard/src/pages/developer-dashboard/tool/ToolsPage.tsx
@@ -1,16 +1,0 @@
-import { useNavigate } from 'react-router';
-import { ToolsList } from '@/components/developer-dashboard/ui/ResourceLists';
-import { Tool } from '@/types/developer-dashboard/appTypes';
-
-export default function ToolsPage() {
-  const navigate = useNavigate();
-
-  return (
-    <ToolsList
-      onCreateClick={() => navigate('/developer/create-tool')}
-      onToolClick={(tool: Tool) =>
-        navigate(`/developer/toolId/${encodeURIComponent(tool.packageName)}`)
-      }
-    />
-  );
-}
