--- packages/apps/app-dashboard/src/pages/developer-dashboard/tool/index.ts
@@ -1,1 +1,0 @@
-export { default as ToolsPage } from './ToolsPage';
