+++ packages/apps/app-dashboard/src/utils/developer-dashboard/formatDateAndTime.ts
@@ -1,0 +1,7 @@
+export const formatDate = (dateString: string) => {
+  return new Date(dateString).toLocaleDateString('en-US', {
+    year: 'numeric',
+    month: 'short',
+    day: 'numeric',
+  });
+};
